Color Correction - fix yellowing (#25)
* Rename whiteBalMatrix and defaultWhiteBalMatrix
Rename whiteBalMatrix to sceneWhiteBalMatrix because it is used in compensating for the color temperature of the lighting in the current scene.
Rename defaultWhiteBalMatrix to cameraWhiteBalMatrix because it is used in compensating for the color temperature of this particular camera.

* Fix white balance turning image yellow

* Fix greenish tint on image before white balance is done.
sceneWhiteBalMatrix values should not just be left at 1, 1, 1 on camera init.
I changed the values for sceneWhiteBalMatrix to be ones that my camera uses for the lighting in my office.  While it may not work as well for other cameras in other lighting, it is far better than just taking the values from cameraWhiteBalance. Also, users who care about color accuracy would do their own white balance anyway instead of using these defaults.

* Rename ccMatrix to colorCalMatrix