Set NoFocus policy on elements of the playback window.
Now that the encoder wheel generates keyPress events, the playback window
needs to receive the keyPress events when open to give a sane playback
experience. To do this, we disable the ability of the buttons on the
playback window from accepting focus so that the playback window will
receive the keyboard events instead.