Revert QWS/Autohide feature. User's don't like it.
Patch 7052af32c29c0228c937cc3374f4eda8fcf08434 added the ability of the camApp
to use the full LCD by autohiding the UI when the user leaves the screen alone
for more than 10 seconds. This seems to be unpopular so we go back a fixed UI
and only 640px of screen for video.