Fixup errors in the calculation of the recorded video regions.
There was a misunderstanding of exactly what the values output from the
recording FIFO meant. Most notably the 'end' is actually the address of
the ending frame of the circular buffer, and 'last' is the address of the
most recently-written frame.

Therefore, the size of the circular buffer is actually computed as:
  recsize = (end - start) + sizeof(frame)
And the starting offset of the first frame should be:
  offset = ((last - start) + sizeof(frame)) % recsize