Fix bug in Save Settings causing the drive would revert to default.
The descriptive flavour text in the drive selection combo box caused
the string match to always fail when searching for the comboBox entry
for the previous drive selection, invoking the default selection to
take effect instead. To fix this, we should compare against mnt.mnt_dir
instead of the comboBox text string.