CamMainWindow: Handle keyPressEvents to control the exposure slider.
Keyboard events from the jog wheel can be used to adjust the exposure
slider as an alternative to using the capacative touch interface. The
Up/Down keys (normal jog wheel) adjust 1/256th of the exposure range
and the PageUp/PageDown keys (pressed jog wheel) adjust by 1/32nd of
the range.