Attempt to handle CCM clipping by scaling instead and remove imgGain.
At higher levels of gain, we find that the imgGain calculation pushed the
color matrix into saturation logic, which badly skews the color space. We
fix this temporarily by removing the imgGain component, and also by
rescaling the matrix in the event that clipping is detected. Ultimately
the imgGain needs to be handled separately.