Add crosshair to whitebalance screen, and remove the popup prompt.
To make it easier to quickly whitebalance the camera in the field, a
crosshair can be drawn on screen to help the user aim the camera at a
white reference point.

Eliminating the popup asking for confirmation also reduces the number
of UI actions to perform a white balance. Accidental presses can still
be undone via the Reset button.