Revise color math now that white balance can occur before demosaic.
The previous workaround to the color saturation problem by doing white
balance in sRGB space gave poor results at extreme color temperatures, but
we can now do the white balance in camera-RGB space so let's fix the math.

We also add a work-in-progress UI window to the WhiteBalance dialog that
allows us to configure the color matrix directly.