Camapp crashes or fails to save if directory not writable
If a directory is not writable for some reason, the camapp would crash when trying to save an h264 file to that directory, or just not save a file at all for the raw16 format. This was caused by how Video::startRecording() compares the value returned from mkfilename() to the enumerated types for errors.