Fix calibration bug leading to exposure mismatch.
If the user adjusts exposure on the camera main window and then attempts
calibration, there was a bug where the exposure time would change, resulting
in a mismatch between the exposure setting after cal. This would also tend
to worsen the calibration quality.

The fix is to ensure that updates to the exposure slider are reflected in
imagerSettings so that the calibration code can pick up the correct exposure.