Move ADC offset training into the sensor class.
We've found that the ADC offset training algorithms don't produce
good quality images on the LUX2100, and by moving the ADC training
algorithm into the sensor class allows the LUX2100 to do something
different vs. the LUX1310.

For the time being, the LUX2100 and LUX2810 just set the ADC offsets
to zero, which seems to produce a good enough image at the expense
of some dynamic range.