Add package upgrade to network screen.
To complete a network update, we must fork and detach a process to
execute 'apt-get upgrade'. The process must be detached since camApp
may be killed during the upgrade. The network upgrade UI still needs
a bunch more polish to give more status back to the user and handle
the cancel buttons correctly.