Pause video during factory cal.
By stopping video playback during factory calibration, we reduce the DDR
memory usage which can help speed up the GPMC frame readout. This also
fixes bug in which the live display was not restarted after factory
calibration.

We also fix a race condition leading to dropped D-Bus messages when sent
between starting the video system and when it becomes available on D-Bus.

Closes: https://github.com/krontech/chronos-cam-app/issues/71
Closes: https://github.com/krontech/chronos-cam-app/issues/68