Round to nearest clock when setting the frame period.
Integer rounding quirks can often lead to incorrect wavetable selection
by choosing the next-smallest wavetable when the frame period happens
to get rounded down below the minimum frame period. Adding half a
clock to the frame period ensures that it rounds to the nearest period
and seems to result in better wavetable selection.

Closes: https://github.com/krontech/chronos-cam-app/issues/64