Remove old IO class and fix bugs in IO1/2 pullups.
There were some bugs in the configuration of the IO1 and IO2 pullups
which were causing the output to always be low (setting the output
to 'none' on an open drain output will pull the output to zero,
defeating the pullups).

We also purge the old IO class from the camApp sources as part of a
cleanup excercise.