Fix: Playback window - Abort Save button is re-enabled before the other buttons #89
Previously, in checkForSaveDone(), camapp was only checking if it was
within 25 frames of the mark out frame. Now, it also considers the
current framerate when deciding whether to disable the Abort button.

Also previously, videoStarted() was only disabling the abort button for
short videos (marked region under 25 frames long). I had to change this
threshold to 230 frames for H.264 format saves because low resolution
videos can be saved at up to about 230fps, which is high enough that
checkForSaveDone() may not get called at all.