Nullify saveDoneTimer pointer to prevent segfaults when saving short videos.
playbackWindow::videoEnded deletes saveDoneTimer, but saveDoneTimer may
be uninitialized if the save is very short(less than 5 frames long).
Ensure that saveDoneTimer is null if it has not yet been initialized,
or if it has already been deleted, so that the if(saveDoneTimer) check
in playbackWindow::videoEnded will actually recognize that the pointer
should not be used.