Fix trigger delay on boot with segmented record mode.
After reboot, camApp would display an incorrect value for post-trigger frames on the trigger delay screen if segmented record mode was used.  This is because, on boot, camApp would set the ratio of post-trigger frames based on the size of the entire length of recording instead of segment size, resulting in a value for triggerPostFrames that was too small by a factor of the number of segments.