Add workaround for max-framerate corruption.
It has been observed that the LUX2100 exhibits frame corruption when
operating close to its maximum framerate. I suspect that it's caused
by the deserializer FIFO in the FPGA not correctly flushing during
the start of a new frame. As a workaround, we can extend the frame
period to give the FIFO time to flush before the end of vblank.