Fix playback -> live state change failure.
There are two bugs that can lead to a missed state change from playback
to live display mode. The first is that the finishedSaving signal can
be emitted from the playback window before executing the destructor,
which can lead to the camera entering recording mode with the playback
window still open.

The second bug is that Camera::setPlayMode() will return an error if
the camera is recording, which can prevent the video system from
returning to live display mode. Instead an error should be returned
only when attempting to go to playback mode while recording.

Closes: https://github.com/krontech/chronos-cam-app/issues/83