Avoid zombie process when restarting app.
The init script that launches camApp appears to kill all instances of
camApp during its startup, which can in turn kill or deadlock the
instance of camApp that is trying to restart itself and this can lead
to a zombie process. Using an exec() call instead of system() to in
this case appears to resolve the issue by renaming the child to
something that killall won't find.

Closes: https://github.com/krontech/chronos-cam-app/issues/98