Update white balance window to use wbTemperature
In the last round of White balance API changes, the wbTemperature
parameter was added to allow the user to configure the white balance
using color temperature (and to allow the sensor driver to provide
sane defaults). However, the GUI was still using the old wbMatrix
API value which no longer has the same effect.

Closes: https://github.com/krontech/chronos-cam-app/issues/114