Combine input and output invert options, and reorganize the IO window.
The reorganization allows us to store the state of the invert and debounce
flags in the ioMappingIo1 and ioMappingIo2 configuration when the function
of the IO is set to output (even though debounce doesn't really make
a lot of sense for an output).

As an extra benefit, this also gives us a bit more room to allow configuring
the isolated input as an exposure input too.

Closes: https://github.com/krontech/chronos-cam-app/issues/122