Always set overlay when loading playback screen.
More recent builds of the video system have added parameters, some of
which are persistent (eg: the overlay). This means that we can't expect
the overlay to always be disabled when opening the playback screen, so
we should therefore always configure it (even if we are just disabling
the overlay).

While we're at it, lets use the parameter API for configuring the
overlay instead of the overlay method call.

Closes: https://github.com/krontech/chronos-cam-app/issues/128