Force use of pullups when frame sync output enabled.
The IO structure of the camera requires a non-zero drive strength
specified when outputting a signal on IO1 or IO2. This was worked
around on the old timing block by implicitly enabling the pullups
whenever the frame-sync output was enabled.

With the new IO block, we have to explictly enable the pullups in
the API. However, the UI should hide this, effectively making the
pullup options meaningless so we add some logic to the UI grey-out
the pullup options whenever frame-sync output is selected.

Closes: https://github.com/krontech/chronos-cam-app/issues/132