Implement settings backup and reversion.
The backup, restoration and reverting-to-default of camera settings
was broken in Debian and API builds for a variety of reasons. Most
notably the majority of the settings are now managed by the API
itself, so simply backing up the QSettings class doesn't do anything.

We also need to update the 'reboot thyself' logic for Debian builds
to work nicely with systemd.

Closes: https://github.com/krontech/chronos-cam-app/issues/129
Closes: https://github.com/krontech/chronos-cam-app/issues/101