Fix eject and format buttons when detecting disks.
The Eject and Format buttons on the storage tab of the util screen
would detect the existance of the first partition on the block device,
rendering them unable to format anything if the disk is unpartitioned.

This is cleaned up simply by using the re-using the status text generated
by lsblk, which already checks for the existence of the block device.

Closes: https://github.com/krontech/chronos-cam-app/issues/148