Cleanup float/int gain conversions.
The gain value reported by the API a floating point value, and generally
reflects the actual gain being applied by the sensor and thus rarely
happens to be a nice integer. However, the GUI uses it more or less as
an enum, so a bit more explicitness is required when converting it.

In paticuar, there are two quirks here: converting a QVariant from a
double to an integer appears to use nearest-integer rounding rather
than floor, and the LUX2100 maximum gain is actually somewhere around
10x rather than 16x as reported by the API.