Clear existing FAT partition when formatting disks.
When (re)formatting a disk via the util window, it is possible that
the new partition table might be the same as the old partition table
in which case the auto-mounter script can mount the drive between
partitioning and formatting and cause the format to fail with an error.

To fix this issue, we wipe the first 1MB of the drive before creating
the partition table to ensure that the start of the new FAT partition
is wiped as well.