Modified LUX8M driver to produce frames of expected size.
The LUX8M driver now produces an image of 4K resolution.
While the vertical alignment of the image seems correct,
the horizontal alignment is off by a fixed amount, which
causes the image to 'wrap' over the horizontal plane.

As the video pipeline can't handle this resolution,
raw frames must be extracted manually.