Cleanup and move digital gain widget to recsettings window.
The digital gain widget was misplaced on the Recording Modes window,
and was badly implemented by mangling the color matrix. With a bit
reshuffling we can move it onto the Recording Settings window, beside
the analog gain widget and we can make it use the digitalGain property
in the API instead.