Use JSON-RPC to update the exposure slider.
Because the exposure slider can move quickly, the GUI can get very laggy
if we have to wait for the D-Bus turnaround time. As a workaround, we can
use the experimental JSON-RPC transport instead and then just do a
send-and-forget for the update.

We also connect the API updates with the exposure slider too so that the
slider will show the correct value even if other users of the API are
making changes.

Closes: https://github.com/krontech/chronos-cam-app/issues/156
Closes: https://github.com/krontech/chronos-cam-app/issues/158