Convert static IP address widget to use nmcli.
By using nmcli, making changes to the IP address configuration can
be done in cooperation with the NetworkManager, which allows things
like cable hotplugging to work as expected.

To make room for the rest of the ancilliary IPv4 details to be set
(netmask and gateway), we had to shrink the SMB configuration window
by combining the server address and share path into one QLineEdit.