Include package download in progress bar motion.
During very large upgrades, if there are a lot of package files to be
downloaded, then the progress bar can sit in one place for a very long
time as the updates are fetched from the internet. To address this
problem we can estimate the number of download steps by using
'apt-get <command> --print-uris --assume-yes' rather than a '--dry-run'.

Closes: https://github.com/krontech/chronos-cam-app/issues/151