camUpdate: Fail gracefully if network is down.
If the network connection is down, there are a couple of problems
that can arise. First, the apt-get update step will fail to donwload
package lists, but will still attempt to hobble by with incomplete
lists. When this occurs, abort the update the procedure and tell
the user to check their network connection.

Furthermore, we should also make an attempt to disable the 'network
upgrade' button until such time as a link is established. This isn't
really a good guarantee that the internet is reachable, but its
better than nothing.

We also fix a hang when attempting to exit rescue mode due to systemd
waiting on the single-user login to complete by sending a SIGHUP to
sulogin. This feels like a bit of a hack, but it's not clear how else
to terminate the single-user login prompt.