Unify refreshDriveList() calls, and fix SMB hang.
By comining the refreshDriveList() functions for both the live
recording mode and the savesettingswindow, we can ensure that they
both get the same set of features.

We also encountered that when a SMB server goes offline, the call
to refresh the drive list can hang while waiting for a network
timeout. This was due to an unnecessary call to stat() the filesystem.