Fix focus issue on Main Screen and White Balance dialog
The background of Main Screen can be highlighted. To fix this issue, remove the button on the background of this dialog.

The Black Cal button is always highlighted (hover with touch screen after pressed). To fix this issue, remove hover effect by pressed effect.

The buttons on White Balance dialog have the same issue with the Black Cal button on Main Screen. Fix it by the same method.

The current selected item in Combo box is highlighted with blue background.

Text color in Status Widgets (Main Screen/ White Balance dialog/ Play Back window) is set to black.