Uncomment two places
Notice there are two places that are commented. One is the trigger delay slider in trigger delay window. It should have red borders to show and set Pre-rec time.

Another is the white balance dialog. After setting the matrix, the crosshair should be shown before setting the matrix button back to <<.