Auto NFS/SMB Connection after reboots
The camera will tryo to connect to the previous NFS (if there are NFS credentials in the appSettings) after reboots. If success, the camera would say anything. If not, a message box would show why the NFS connection failed.

If NFS connection fails and there re crendentials fro SMB, the camera will try to connect to SMB after users click OK button in the NFS Connection Sttaus message box.